\name{geosapi-package}
\alias{geosapi-package}
\alias{geosapi}
\docType{package}
\title{
\packageTitle{geosapi}
}
\description{
\packageDescription{geosapi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{geosapi}
\packageIndices{geosapi}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{geosapi}

Maintainer: \packageMaintainer{geosapi}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
