% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosapi.R
\docType{package}
\name{geosapi}
\alias{geosapi}
\alias{geosapi-package}
\title{GeoServer REST API R Interface}
\description{
Provides an R interface to the GeoServer REST API, allowing to 
upload and publish data in a GeoServer web-application and expose data to OGC 
Web-Services. The package currently supports all CRUD (Create,Read,Update,Delete) 
operations on GeoServer workspaces, namespaces, datastores (stores of vector data), 
featuretypes, layers, styles, as well as vector data upload operations. For 
more information about the GeoServer REST API, see <https://docs.geoserver.org/main/en/user/rest/> .
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eblondel/geosapi}
  \item \url{https://eblondel.github.io/geosapi/}
  \item \url{https://geoserver.org/}
  \item Report bugs at \url{https://github.com/eblondel/geosapi/issues}
}

}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
