% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAbstractCoverageStore.R
\docType{class}
\name{GSAbstractCoverageStore}
\alias{GSAbstractCoverageStore}
\title{Geoserver REST API CoverageStore}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer CoverageStore
}
\description{
Geoserver REST API CoverageStore
}
\section{Methods}{

\describe{
 \item{\code{new(xml, type, name, description, enabled)}}{
   This method is used to instantiate a \code{GSAbstractCoverageStore}
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a \code{GSAbstractCoverageStore} from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a \code{GSAbstractCoverageStore} to XML. Inherited from the
   generic \code{GSRESTResource} encoder
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CoverageStore}
\keyword{api}
\keyword{geoserver}
\keyword{rest}
