% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAbstractStore.R
\docType{class}
\name{GSAbstractStore}
\alias{GSAbstractStore}
\title{Geoserver REST API Store}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer store
}
\description{
Geoserver REST API Store
}
\section{Fields}{

\describe{
\item{\code{name}}{store name}

\item{\code{description}}{store description}

\item{\code{enabled}}{if the store is enabled or not}

\item{\code{type}}{store type}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, storeType, type, name, description, enabled)}}{
   This method is used to instantiate a \code{GSAbstractStore}
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a \code{GSAbstractStore} from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a \code{GSAbstractStore} to XML. Inherited from the
   generic \code{GSRESTResource} encoder
 }
 \item{\code{setEnabled(enabled)}}{
   Sets the datastore as enabled if TRUE, disabled if FALSE
 }
 \item{\code{setDescription(description)}}{
   Sets the datastore description
 }
 \item{\code{setType(type)}}{
   Sets the datastore type
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{rest}
\keyword{store}
