% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSArcGridCoverageStore.R
\docType{class}
\name{GSArcGridCoverageStore}
\alias{GSArcGridCoverageStore}
\title{Geoserver REST API ArcGridCoverageStore}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer ArcGrid CoverageStore
}
\description{
Geoserver REST API ArcGridCoverageStore
}
\section{Methods}{

\describe{
 \item{\code{new(xml, name, description, enabled, url)}}{
   This method is used to instantiate a \code{GSArcGridCoverageStore}
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ArcGrid}
\keyword{CoverageStore}
\keyword{api}
\keyword{geoserver}
\keyword{rest}
