% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSCoverage.R
\docType{class}
\name{GSCoverage}
\alias{GSCoverage}
\title{A GeoServer coverage}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer coverage
}
\description{
This class models a GeoServer coverage. This class is to be
used for manipulating representations of vector data with GeoServer.
}
\details{
Geoserver REST API Resource
}
\section{Methods}{

\describe{
 \item{\code{new(rootName, xml)}}{
   This method is used to instantiate a \code{GSCoverage}
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a \code{GSCoverage}. Inherited from the
   generic \code{GSRESTResource} encoder
 }
 \item{\code{encode()}}{
   This method is used to encode a \code{GSCoverage} to XML. Inherited from the
   generic \code{GSRESTResource} encoder
 }
 \item{\code{setView(cv)}}{
   Sets a coverage view for the coverage.
 }
 \item{\code{delView()}}{
   Deletes the coverage view for the coverage
 }
}
}

\examples{
gt <- GSCoverage$new()

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{coverageType}
\keyword{geoserver}
\keyword{resource}
\keyword{rest}
