% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSCoverageBand.R
\docType{class}
\name{GSCoverageBand}
\alias{GSCoverageBand}
\title{Geoserver REST API GSCoverageBand}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer coverage band
}
\description{
Geoserver REST API GSCoverageBand
}
\section{Fields}{

\describe{
\item{\code{inputCoverageBands}}{list of input coverage bands}

\item{\code{definition}}{coverage band definition}

\item{\code{index}}{coverage band index}

\item{\code{compositionType}}{coverage band composition type}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate a \code{GSCoverageBand}
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a \code{GSCoverageBand} from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a \code{GSCoverageBand} to XML
 }
 \item{\code{setDefinition(definition)}}{
   Sets the coverage band definition
 }
 \item{\code{setIndex(index)}}{
   Sets the coverage band index
 }
 \item{\code{setCompositionType}}{
   Sets the composition type. Only 'BAND_SELECT' is supported by GeoServer for now.
 }
 \item{\code{addInputBand(band)}}{
   Adds a input coverage band, object of class \code{GSInputCoverageBand}
 }
 \item{\code{delInputBand(band)}}{
   Removes a input coverage band, object of class \code{GSInputCoverageBand}
 }
}
}

\examples{
GSCoverageBand$new()

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{coverageBand}
\keyword{geoserver}
\keyword{rest}
