% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSCoverageView.R
\docType{class}
\name{GSCoverageView}
\alias{GSCoverageView}
\title{Geoserver REST API GSCoverageView}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer coverage view
}
\description{
Geoserver REST API GSCoverageView
}
\section{Fields}{

\describe{
\item{\code{name}}{name}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate a \code{GSCoverageView}
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a \code{GSCoverageView} from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a \code{GSCoverageView} to XML
 }
 \item{\code{setName(name)}}{
   Sets the name of the coverage view
 }
 \item{\code{setEnvelopeCompositionType}}{
   Sets the envelope composition type. Type of Envelope Composition, used to expose the bounding box 
   of the CoverageView, either 'UNION' or 'INTERSECTION'.
 }
 \item{\code{setSelectedResolution(selectedResolution)}}{
   Sets the selected resolution
 }
 \item{\code{setSelectedResolutionIndex(selectedResolutionIndex)}}{
   Sets the selected resolution index
 }
 \item{\code{addBand(band)}}{
   Adds a coverage band
 }
 \item{\code{delBand(band)}}{
   Removes a coverage band.
 }
}
}

\examples{
GSCoverageView$new()

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{coverageView}
\keyword{geoserver}
\keyword{rest}
