% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSDataStore.R
\docType{class}
\name{GSDataStore}
\alias{GSDataStore}
\title{Geoserver REST API DataStore}
\format{\code{\link{R6Class}} object.}
\usage{
GSDataStore
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer dataStore
}
\description{
Geoserver REST API DataStore
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{workspace}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate a GSDataStoreManager
 }
}
}
\examples{
GSDataStore$new(xml)
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{DataStore}
\keyword{api}
\keyword{geoserver}
\keyword{rest}

