% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSFeatureDimension.R
\docType{class}
\name{GSFeatureDimension}
\alias{GSFeatureDimension}
\title{A GeoServer dimension}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer feature dimension
}
\description{
This class models a GeoServer feature dimension.
}
\examples{
  dim <- GSFeatureDimension$new()

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{dimension}
\keyword{geoserver}
\keyword{resource}
\keyword{rest}
\section{Super classes}{
\code{\link[geosapi:GSRESTResource]{geosapi::GSRESTResource}} -> \code{\link[geosapi:GSDimension]{geosapi::GSDimension}} -> \code{GSFeatureDimension}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attribute}}{attribute}

\item{\code{endAttribute}}{end attribute}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GSFeatureDimension-new}{\code{GSFeatureDimension$new()}}
\item \href{#method-GSFeatureDimension-decode}{\code{GSFeatureDimension$decode()}}
\item \href{#method-GSFeatureDimension-setAttribute}{\code{GSFeatureDimension$setAttribute()}}
\item \href{#method-GSFeatureDimension-setEndAttribute}{\code{GSFeatureDimension$setEndAttribute()}}
\item \href{#method-GSFeatureDimension-clone}{\code{GSFeatureDimension$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSRESTResource" data-id="encode"><a href='../../geosapi/html/GSRESTResource.html#method-GSRESTResource-encode'><code>geosapi::GSRESTResource$encode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSDimension" data-id="setEnabled"><a href='../../geosapi/html/GSDimension.html#method-GSDimension-setEnabled'><code>geosapi::GSDimension$setEnabled()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSDimension" data-id="setPresentation"><a href='../../geosapi/html/GSDimension.html#method-GSDimension-setPresentation'><code>geosapi::GSDimension$setPresentation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSDimension" data-id="setUnit"><a href='../../geosapi/html/GSDimension.html#method-GSDimension-setUnit'><code>geosapi::GSDimension$setUnit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSDimension" data-id="setUnitSymbol"><a href='../../geosapi/html/GSDimension.html#method-GSDimension-setUnitSymbol'><code>geosapi::GSDimension$setUnitSymbol()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSFeatureDimension-new"></a>}}
\if{latex}{\out{\hypertarget{method-GSFeatureDimension-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object of class \link{GSFeatureDimension}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSFeatureDimension$new(xml = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{XMLInternalNode-class}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSFeatureDimension-decode"></a>}}
\if{latex}{\out{\hypertarget{method-GSFeatureDimension-decode}{}}}
\subsection{Method \code{decode()}}{
Decodes from XML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSFeatureDimension$decode(xml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{XMLInternalNode-class}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSFeatureDimension-setAttribute"></a>}}
\if{latex}{\out{\hypertarget{method-GSFeatureDimension-setAttribute}{}}}
\subsection{Method \code{setAttribute()}}{
Set attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSFeatureDimension$setAttribute(attribute)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attribute}}{attribute}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSFeatureDimension-setEndAttribute"></a>}}
\if{latex}{\out{\hypertarget{method-GSFeatureDimension-setEndAttribute}{}}}
\subsection{Method \code{setEndAttribute()}}{
Set end attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSFeatureDimension$setEndAttribute(endAttribute)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endAttribute}}{end attribute}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSFeatureDimension-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GSFeatureDimension-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSFeatureDimension$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
