% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSGeoPackageDataStore.R
\docType{class}
\name{GSGeoPackageDataStore}
\alias{GSGeoPackageDataStore}
\title{Geoserver REST API GeoPackageDataStore}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer GeoPackage dataStore
}
\description{
Geoserver REST API GeoPackageDataStore
}
\section{Methods}{

\describe{
   \item{\code{new(xml, dataStore, description, enabled, url)}}{
     Instantiates a GSGeoPackageDataStore object
   }
   \item{\code{setUrl(url)}}{
     Set the spatial files data URL
   }
}
}

\examples{
GSGeoPackageDataStore$new(dataStore="ds", description = "des",
                         enabled = TRUE, url = "file://data/file.gpkg")

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{DataStore}
\keyword{GeoPackage}
\keyword{api}
\keyword{geoserver}
\keyword{rest}
