% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSGeoTIFFCoverageStore.R
\docType{class}
\name{GSGeoTIFFCoverageStore}
\alias{GSGeoTIFFCoverageStore}
\title{Geoserver REST API GeoTIFF CoverageStore}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer GeoTIFF CoverageStore
}
\description{
Geoserver REST API GeoTIFF CoverageStore
}
\section{Methods}{

\describe{
 \item{\code{new(xml, name, description, enabled, url)}}{
   This method is used to instantiate a \code{GSGeoTIFFCoverageStore}
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CoverageStore}
\keyword{GeoTIFF}
\keyword{api}
\keyword{geoserver}
\keyword{rest}
