% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSImageMosaicCoverageStore.R
\docType{class}
\name{GSImageMosaicCoverageStore}
\alias{GSImageMosaicCoverageStore}
\title{Geoserver REST API ImageMosaicCoverageStore}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer ImageMosaic CoverageStore
}
\description{
Geoserver REST API ImageMosaicCoverageStore
}
\section{Methods}{

\describe{
 \item{\code{new(xml, name, description, enabled, url)}}{
   This method is used to instantiate a \code{GSImageMosaicCoverageStore}
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CoverageStore}
\keyword{ImageMosaic}
\keyword{api}
\keyword{geoserver}
\keyword{rest}
