% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSInputCoverageBand.R
\docType{class}
\name{GSInputCoverageBand}
\alias{GSInputCoverageBand}
\title{Geoserver REST API GSInputCoverageBand}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer input coverage band
}
\description{
Geoserver REST API GSInputCoverageBand
}
\section{Fields}{

\describe{
\item{\code{coverageName}}{coverage name}

\item{\code{band}}{coverage band}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, coverageName, band)}}{
   This method is used to instantiate a \code{GSInputCoverageBand}
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a \code{GSInputCoverageBand} from XML
 }
 \item{\code{setCoverageName(coverageName)}}{
   Sets the coverage name
 }
 \item{\code{setBand(band)}}{
   Sets the coverage band
 }
}
}

\examples{
GSInputCoverageBand$new()

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{inputCoverageBand}
\keyword{rest}
