% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSLayer.R
\docType{class}
\name{GSLayer}
\alias{GSLayer}
\title{A GeoServer layer resource}
\format{\code{\link{R6Class}} object.}
\usage{
GSLayer
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer layer
}
\description{
This class models a GeoServer layer. This class is to be
used for published resource (feature type or coverage).
}
\details{
Geoserver REST API Resource
}
\section{Methods}{

\describe{
 \item{\code{new(rootName, xml)}}{
   This method is used to instantiate a GSLayer
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a GSLayer from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a GSLayer to XML. Inherited from the
   generic \code{GSRESTResource} encoder
 }
}
}
\examples{
lyr <- GSLayer$new()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{coverage}
\keyword{featureType}
\keyword{geoserver}
\keyword{layer}
\keyword{resource}
\keyword{rest}

