% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSMetadataLink.R
\docType{class}
\name{GSMetadataLink}
\alias{GSMetadataLink}
\title{A GeoServer resource metadataLink}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer resource metadataLink
}
\description{
This class models a GeoServer resource metadataLink made of a type
(free text e.g. text/xml, text/html), a metadataType (Possible values are 
ISO19115:2003, FGDC, TC211, 19139, other), and a content: an URL that gives 
the metadataLink
}
\details{
Geoserver REST API Metadatalink
}
\section{Methods}{

\describe{
 \item{\code{new(xml, type, metadataType, content)}}{
   This method is used to instantiate a GSMetadataLink
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a GSMetadataLink from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a GSMetadataLink to XML. Inherited from the
   generic \code{GSRESTResource} encoder
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{metadataLink}
\keyword{resource}
\keyword{rest}
