% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSPublishable.R
\docType{class}
\name{GSPublishable}
\alias{GSPublishable}
\title{A GeoServer layer group publishable}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer layer group publishable
}
\description{
This class models a GeoServer layer. This class is to be
used internally by \pkg{geosapi} for configuring layers or layer groups
within an object of class \code{GSLayerGroup}
}
\examples{
  publishable <- GSPublishable$new(name = "name", type = "layer")

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{group}
\keyword{layer}
\keyword{publishable}
\keyword{resource}
\keyword{rest}
\section{Super class}{
\code{\link[geosapi:GSRESTResource]{geosapi::GSRESTResource}} -> \code{GSPublishable}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{full}}{full}

\item{\code{name}}{name}

\item{\code{attr_type}}{type of attribute}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GSPublishable-new}{\code{GSPublishable$new()}}
\item \href{#method-GSPublishable-decode}{\code{GSPublishable$decode()}}
\item \href{#method-GSPublishable-setName}{\code{GSPublishable$setName()}}
\item \href{#method-GSPublishable-setType}{\code{GSPublishable$setType()}}
\item \href{#method-GSPublishable-clone}{\code{GSPublishable$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSRESTResource" data-id="encode"><a href='../../geosapi/html/GSRESTResource.html#method-GSRESTResource-encode'><code>geosapi::GSRESTResource$encode()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSPublishable-new"></a>}}
\if{latex}{\out{\hypertarget{method-GSPublishable-new}{}}}
\subsection{Method \code{new()}}{
Initializes a \link{GSPublishable}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSPublishable$new(xml = NULL, name, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{an object of class \link{XMLInternalNode-class}}

\item{\code{name}}{name}

\item{\code{type}}{type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSPublishable-decode"></a>}}
\if{latex}{\out{\hypertarget{method-GSPublishable-decode}{}}}
\subsection{Method \code{decode()}}{
Decodes from XML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSPublishable$decode(xml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{an object of class \link{XMLInternalNode-class}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSPublishable-setName"></a>}}
\if{latex}{\out{\hypertarget{method-GSPublishable-setName}{}}}
\subsection{Method \code{setName()}}{
set name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSPublishable$setName(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSPublishable-setType"></a>}}
\if{latex}{\out{\hypertarget{method-GSPublishable-setType}{}}}
\subsection{Method \code{setType()}}{
Set type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSPublishable$setType(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSPublishable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GSPublishable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSPublishable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
