% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSRule.R
\docType{class}
\name{GSRule}
\alias{GSRule}
\title{A GeoServer access control list rule}
\format{
\code{\link[R6]{R6Class}} object.
}
\value{
Object of \code{\link[R6]{R6Class}} for modelling a GeoServer access control list rule
}
\description{
This class models a GeoServer access control list rule
}
\details{
Geoserver REST API Access Control List Rule
}
\note{
Abstract class
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{access}
\keyword{api}
\keyword{control}
\keyword{geoserver}
\keyword{rest}
\keyword{rule}
\section{Super class}{
\code{\link[geosapi:GSRESTResource]{geosapi::GSRESTResource}} -> \code{GSRule}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GSRule-new}{\code{GSRule$new()}}
\item \href{#method-GSRule-encode}{\code{GSRule$encode()}}
\item \href{#method-GSRule-decode}{\code{GSRule$decode()}}
\item \href{#method-GSRule-clone}{\code{GSRule$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSRESTResource" data-id="getClassName"><a href='../../geosapi/html/GSRESTResource.html#method-GSRESTResource-getClassName'><code>geosapi::GSRESTResource$getClassName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSRESTResource" data-id="print"><a href='../../geosapi/html/GSRESTResource.html#method-GSRESTResource-print'><code>geosapi::GSRESTResource$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRule-new"></a>}}
\if{latex}{\out{\hypertarget{method-GSRule-new}{}}}
\subsection{Method \code{new()}}{
Initializes a \link{GSRule}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRule$new(xml = NULL, domain = c("layers", "services", "rest"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{an object of class \link[xml2]{xml_node-class}}

\item{\code{domain}}{the access control domain}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRule-encode"></a>}}
\if{latex}{\out{\hypertarget{method-GSRule-encode}{}}}
\subsection{Method \code{encode()}}{
Encodes as XML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRule$encode()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \link[xml2]{xml_node-class}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRule-decode"></a>}}
\if{latex}{\out{\hypertarget{method-GSRule-decode}{}}}
\subsection{Method \code{decode()}}{
Decodes from XML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRule$decode(xml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{an object of class \link[xml2]{xml_node-class}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GSRule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
