% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSShinyMonitor.R
\docType{class}
\name{GSShinyMonitor}
\alias{GSShinyMonitor}
\title{Geoserver REST API DataStore}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for setting a GS Shiny monitoring app
}
\description{
Geoserver REST API DataStore
}
\note{
Internal class used for \code{GSManager$monitor} method
}
\section{Methods}{

\describe{
 \item{\code{new(manager, sleep)}}{
   This method is used to instantiate a R shiny monitoring app for a Geoserver 
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{geoserver}
\keyword{monitoring}
