% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSVirtualTable.R
\docType{class}
\name{GSVirtualTable}
\alias{GSVirtualTable}
\title{Geoserver REST API GSVirtualTable}
\format{\code{\link{R6Class}} object.}
\usage{
GSVirtualTable
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer virtual table
}
\description{
Geoserver REST API GSVirtualTable
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{sql}}{}

\item{\code{escapeSql}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate a GSVirtualTable
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a GSVirtualTable from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a GSVirtualTable to XML
 }
}
}
\examples{
GSVirtualTable$new()
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{rest}
\keyword{virtualTable}

