% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSWorkspace.R
\docType{class}
\name{GSWorkspace}
\alias{GSWorkspace}
\title{Geoserver REST API Workspace}
\format{\code{\link{R6Class}} object.}
\usage{
GSWorkspace
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer workspace
}
\description{
Geoserver REST API Workspace
}
\section{Fields}{

\describe{
\item{\code{xml}}{}

\item{\code{name}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate a GSWorkspace
 }
}
}
\examples{
GSWorkspace$new(xml)
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{rest}
\keyword{workspace}

