% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSWorldImageCoverageStore.R
\docType{class}
\name{GSWorldImageCoverageStore}
\alias{GSWorldImageCoverageStore}
\title{Geoserver REST API WorldImageCoverageStore}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer WorldImage CoverageStore
}
\description{
Geoserver REST API WorldImageCoverageStore
}
\section{Methods}{

\describe{
 \item{\code{new(xml, name, description, enabled, url)}}{
   This method is used to instantiate a \code{GSWorldImageCoverageStore}
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{CoverageStore}
\keyword{WorldImage}
\keyword{api}
\keyword{geoserver}
\keyword{rest}
