% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geosapi.R
\docType{package}
\name{geosapi}
\alias{geosapi}
\alias{geosapi-package}
\title{Geoserver REST API R Interface}
\description{
Provides an R interface to the Geoserver REST API, allowing to
upload and publish data in a GeoServer web-application and expose data to OGC
Web-Services. The package currently supports all CRUD (Create,Read,Update,Delete)
operations on GeoServer workspaces, namespaces, datastores (stores of vector data),
featuretypes, layers, styles, as well as vector data upload operations. For
more information about the GeoServer REST API, see http://docs.geoserver.org/stable/en/user/rest/ .
}
\details{
\tabular{ll}{
   Package: \tab geosapi\cr
   Type: \tab Package\cr
   Version: \tab 0.1-0\cr
   Date: \tab 2017-02-21\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}

