% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geosapi.R
\docType{package}
\name{geosapi}
\alias{geosapi}
\alias{geosapi-package}
\title{Geoserver REST API R Interface}
\description{
geosapi provides an R interface to the GeoServer REST web-services API.
}
\details{
\tabular{ll}{
   Package: \tab geosapi\cr
   Type: \tab Package\cr
   Version: \tab 0.1\cr
   Date: \tab 2017-01-31\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}

