% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSDatastoreManager.R
\docType{class}
\name{GSDatastoreManager}
\alias{GSDatastoreManager}
\title{Geoserver REST API Datastore Manager}
\format{\code{\link{R6Class}} object.}
\usage{
GSDatastoreManager
}
\value{
Object of \code{\link{R6Class}} with methods for managing GeoServer
datastores (i.e. stores of vector data)
}
\description{
Geoserver REST API Datastore Manager
}
\section{Methods}{

\describe{
 \item{\code{new(}}{
   This method is used to instantiate a GSDatastoreManager
 }
 \item{\code{uploadData(workspace, datastore, endpoint, extension,
                        configure, update, filename, charset, contentType)}}{
   Uploads data to a target datastore
 }
 \item{\code{uploadShapefile(workspace, datastore, endpoint,
                             configure, update, filename, charset)}}{
   Uploads a zipped ESRIshapefile to a target datastore
 }
 \item{\code{uploadProperties(workspace, datastore, endpoint,
                             configure, update, filename, charset)}}{
   Uploads a properties file to a target datastore
 }
 \item{\code{uploadH2(workspace, datastore, endpoint,
                             configure, update, filename, charset)}}{
   Uploads a H2 database to a target datastore
 }
 \item{\code{uploadSpatialite(workspace, datastore, endpoint,
                             configure, update, filename, charset)}}{
   Uploads a Spatialite database to a target datastore
 }
 \item{\code{uploadAppschema(workspace, datastore, endpoint,
                             configure, update, filename, charset)}}{
   Uploads a appschema file to a target datastore
 }
}
}
\examples{
GSDatastoreManager$new("http://localhost:8080/geoserver", "admin", "geoserver")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{datastore}
\keyword{geoserver}
\keyword{rest}

