% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDimension.R, R/GSFeatureDimension.R
\docType{class}
\name{GSDimension}
\alias{GSDimension}
\alias{GSFeatureDimension}
\title{A GeoServer dimension}
\format{
\code{\link{R6Class}} object.

\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer dimension

Object of \code{\link{R6Class}} for modelling a GeoServer feature dimension
}
\description{
This class models a GeoServer resource dimension.

This class models a GeoServer feature dimension.
}
\details{
Geoserver REST API Dimension

Geoserver REST API FeatureDimension
}
\section{Fields}{

\describe{
\item{\code{enabled}}{true/false}

\item{\code{presentation}}{dimension presentation}

\item{\code{resolution}}{dimension resolution}

\item{\code{units}}{dimension units}

\item{\code{unitSymbol}}{dimension unitsSymbol}

\item{\code{attribute}}{attribute}

\item{\code{endAttribute}}{end attribute}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate a GSResource
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a GSResource from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a GSFeatureType to XML. Inherited from the
   generic \code{GSRESTResource} encoder
 }
}


\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate a GSResource
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a GSResource from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a GSFeatureType to XML. Inherited from the
   generic \code{GSRESTResource} encoder
 }
}
}

\examples{
dim <- GSDimension$new()

dim <- GSFeatureDimension$new()

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{dimension}
\keyword{geoserver}
\keyword{resource}
\keyword{rest}
