% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSFeatureDimension.R
\docType{class}
\name{GSFeatureDimension}
\alias{GSFeatureDimension}
\title{A GeoServer dimension}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer feature dimension
}
\description{
This class models a GeoServer feature dimension.
}
\examples{
  dim <- GSFeatureDimension$new()

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{dimension}
\keyword{geoserver}
\keyword{resource}
\keyword{rest}
\section{Super classes}{
\code{\link[geosapi:GSRESTResource]{geosapi::GSRESTResource}} -> \code{\link[geosapi:GSDimension]{geosapi::GSDimension}} -> \code{GSFeatureDimension}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{attribute}}{attribute}

\item{\code{endAttribute}}{end attribute}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GSFeatureDimension$new()}}
\item \href{#method-decode}{\code{GSFeatureDimension$decode()}}
\item \href{#method-setAttribute}{\code{GSFeatureDimension$setAttribute()}}
\item \href{#method-setEndAttribute}{\code{GSFeatureDimension$setEndAttribute()}}
\item \href{#method-clone}{\code{GSFeatureDimension$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="geosapi" data-topic="GSRESTResource" data-id="encode">}\href{../../geosapi/html/GSRESTResource.html#method-encode}{\code{geosapi::GSRESTResource$encode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geosapi" data-topic="GSDimension" data-id="setEnabled">}\href{../../geosapi/html/GSDimension.html#method-setEnabled}{\code{geosapi::GSDimension$setEnabled()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geosapi" data-topic="GSDimension" data-id="setPresentation">}\href{../../geosapi/html/GSDimension.html#method-setPresentation}{\code{geosapi::GSDimension$setPresentation()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geosapi" data-topic="GSDimension" data-id="setUnit">}\href{../../geosapi/html/GSDimension.html#method-setUnit}{\code{geosapi::GSDimension$setUnit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="geosapi" data-topic="GSDimension" data-id="setUnitSymbol">}\href{../../geosapi/html/GSDimension.html#method-setUnitSymbol}{\code{geosapi::GSDimension$setUnitSymbol()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object of class \link{GSFeatureDimension}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSFeatureDimension$new(xml = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{XMLInternalNode-class}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-decode"></a>}}
\if{latex}{\out{\hypertarget{method-decode}{}}}
\subsection{Method \code{decode()}}{
Decodes from XML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSFeatureDimension$decode(xml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{XMLInternalNode-class}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setAttribute"></a>}}
\if{latex}{\out{\hypertarget{method-setAttribute}{}}}
\subsection{Method \code{setAttribute()}}{
Set attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSFeatureDimension$setAttribute(attribute)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{attribute}}{attribute}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setEndAttribute"></a>}}
\if{latex}{\out{\hypertarget{method-setEndAttribute}{}}}
\subsection{Method \code{setEndAttribute()}}{
Set end attribute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSFeatureDimension$setEndAttribute(endAttribute)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endAttribute}}{end attribute}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSFeatureDimension$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
