% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSInputCoverageBand.R
\docType{class}
\name{GSInputCoverageBand}
\alias{GSInputCoverageBand}
\title{Geoserver REST API GSInputCoverageBand}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer input coverage band
}
\description{
Geoserver REST API GSInputCoverageBand

Geoserver REST API GSInputCoverageBand
}
\section{Methods}{

\describe{
 \item{\code{new(xml, coverageName, band)}}{
   This method is used to instantiate a \code{GSInputCoverageBand}
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a \code{GSInputCoverageBand} from XML
 }
 \item{\code{setCoverageName(coverageName)}}{
   Sets the coverage name
 }
 \item{\code{setBand(band)}}{
   Sets the coverage band
 }
}
}

\examples{
  GSInputCoverageBand$new()

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{inputCoverageBand}
\keyword{rest}
\section{Super class}{
\code{\link[geosapi:GSRESTResource]{geosapi::GSRESTResource}} -> \code{GSInputCoverageBand}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{coverageName}}{coverage name}

\item{\code{band}}{band}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GSInputCoverageBand-new}{\code{GSInputCoverageBand$new()}}
\item \href{#method-GSInputCoverageBand-decode}{\code{GSInputCoverageBand$decode()}}
\item \href{#method-GSInputCoverageBand-setCoverageName}{\code{GSInputCoverageBand$setCoverageName()}}
\item \href{#method-GSInputCoverageBand-setBand}{\code{GSInputCoverageBand$setBand()}}
\item \href{#method-GSInputCoverageBand-clone}{\code{GSInputCoverageBand$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSRESTResource" data-id="encode"><a href='../../geosapi/html/GSRESTResource.html#method-GSRESTResource-encode'><code>geosapi::GSRESTResource$encode()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSInputCoverageBand-new"></a>}}
\if{latex}{\out{\hypertarget{method-GSInputCoverageBand-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object of class \link{GSInputCoverageBand}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSInputCoverageBand$new(xml = NULL, coverageName = NULL, band = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{XMLInternalNode-class}}

\item{\code{coverageName}}{coverage name}

\item{\code{band}}{band name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSInputCoverageBand-decode"></a>}}
\if{latex}{\out{\hypertarget{method-GSInputCoverageBand-decode}{}}}
\subsection{Method \code{decode()}}{
Decodes from XML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSInputCoverageBand$decode(xml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{XMLInternalNode-class}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSInputCoverageBand-setCoverageName"></a>}}
\if{latex}{\out{\hypertarget{method-GSInputCoverageBand-setCoverageName}{}}}
\subsection{Method \code{setCoverageName()}}{
Set coverage name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSInputCoverageBand$setCoverageName(coverageName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coverageName}}{coverage name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSInputCoverageBand-setBand"></a>}}
\if{latex}{\out{\hypertarget{method-GSInputCoverageBand-setBand}{}}}
\subsection{Method \code{setBand()}}{
Set band
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSInputCoverageBand$setBand(band)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{band}}{band}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSInputCoverageBand-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GSInputCoverageBand-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSInputCoverageBand$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
