% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSManager.R
\docType{class}
\name{GSManager}
\alias{GSManager}
\title{Geoserver REST API Manager}
\format{\code{\link{R6Class}} object.}
\usage{
GSManager
}
\value{
Object of \code{\link{R6Class}} with methods for communication with
the REST API of a GeoServer instance.
}
\description{
Geoserver REST API Manager
}
\section{Fields}{

\describe{
\item{\code{baseUrl}}{the base URL of the GeoServer REST API}

\item{\code{token}}{the authentication token used to connect to Geoserver}
}}
\section{Methods}{

\describe{
 \item{\code{new(url, user, pwd)}}{
   This method is used to instantiate a GSManager with the \code{url} of the
   GeoServer and credentials to authenticate (\code{user}/\code{pwd})
 }
 \item{\code{getBaseUrl()}}{
   This method gives the base URL of the GeoServer REST API.
 }
 \item{\code{getToken()}}{
   This method gives the authentication token to connect to Geoserver
 }
 \item{\code{GET(path)}}{
   This method performs a GET request for a given \code{path} to GeoServer REST API
 }
 \item{\code{PUT(path, filename, contentType)}}{
   This method performs a PUT request for a given \code{path} to GeoServer REST API,
   to upload a file of name \code{filename} with given \code{contentType}
 }
 \item{\code{connect()}}{
   This methods attempts a connection to GeoServer REST API. User internally
   during initialization of \code{GSManager}.
 }
 \item{\code{getClassName()}}{
   Retrieves the name of the class instance
 }
}
}
\examples{
GSManager$new("http://localhost:8080/geoserver", "admin", "geoserver")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{rest}

