% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSManager.R
\docType{class}
\name{GSManager}
\alias{GSManager}
\title{Geoserver REST API Manager}
\format{\code{\link{R6Class}} object.}
\usage{
GSManager
}
\value{
Object of \code{\link{R6Class}} with methods for communication with
the REST API of a GeoServer instance.
}
\description{
Geoserver REST API Manager
}
\section{Fields}{

\describe{
\item{\code{baseUrl}}{the Base url of GeoServer}

\item{\code{verbose}}{if logs have to be printed}
}}
\section{Methods}{

\describe{
 \item{\code{new(url, user, pwd)}}{
   This method is used to instantiate a GSManager with the \code{url} of the
   GeoServer and credentials to authenticate (\code{user}/\code{pwd})
 }
 \item{\code{getUrl()}}{
   Get the authentication URL
 }
 \item{\code{connect()}}{
   This methods attempts a connection to GeoServer REST API. User internally
   during initialization of \code{GSManager}.
 }
 \item{\code{getClassName()}}{
   Retrieves the name of the class instance
 }
 \item{\code{getWorkspaceManager()}}{
   Retrieves an instance of workspace manager
 }
 \item{\code{getNamespaceManager()}}{
   Retrieves an instance of namespace manager
 }
 \item{\code{getDataStoreManager()}}{
   Retrieves an instance of datastore manager
 }
}
}
\examples{
\dontrun{
   GSManager$new("http://localhost:8080/geoserver", "admin", "geoserver")
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{rest}

