% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSNamespace.R
\docType{class}
\name{GSNamespace}
\alias{GSNamespace}
\title{Geoserver REST API Namespace}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer namespace
}
\description{
Geoserver REST API Namespace
}
\section{Fields}{

\describe{
\item{\code{name}}{namespace name}

\item{\code{prefix}}{namespace prefix}

\item{\code{uri}}{namespace URI}

\item{\code{full}}{completeness of the namespace description}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, prefix, uri)}}{
   This method is used to instantiate a GSNamespace
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a GSNamespace from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a GSNamespace to XML. Inherited from the
   generic \code{GSRESTResource} encoder
 }
}
}

\examples{
GSNamespace$new(prefix = "prefix", uri = "http://prefix")

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{namespace}
\keyword{rest}
