% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSNamespace.R
\docType{class}
\name{GSNamespace}
\alias{GSNamespace}
\title{Geoserver REST API Namespace}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer namespace
}
\description{
Geoserver REST API Namespace

Geoserver REST API Namespace
}
\examples{
GSNamespace$new(prefix = "my_ns", uri = "http://my_ns")

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{namespace}
\keyword{rest}
\section{Super class}{
\code{\link[geosapi:GSRESTResource]{geosapi::GSRESTResource}} -> \code{GSNamespace}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{namespace name}

\item{\code{prefix}}{namespace prefix}

\item{\code{uri}}{namespace URI}

\item{\code{full}}{completeness of the namespace description}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GSNamespace-new}{\code{GSNamespace$new()}}
\item \href{#method-GSNamespace-decode}{\code{GSNamespace$decode()}}
\item \href{#method-GSNamespace-clone}{\code{GSNamespace$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSRESTResource" data-id="encode"><a href='../../geosapi/html/GSRESTResource.html#method-GSRESTResource-encode'><code>geosapi::GSRESTResource$encode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSRESTResource" data-id="getClassName"><a href='../../geosapi/html/GSRESTResource.html#method-GSRESTResource-getClassName'><code>geosapi::GSRESTResource$getClassName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSRESTResource" data-id="print"><a href='../../geosapi/html/GSRESTResource.html#method-GSRESTResource-print'><code>geosapi::GSRESTResource$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSNamespace-new"></a>}}
\if{latex}{\out{\hypertarget{method-GSNamespace-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object of class \link{GSNamespace}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSNamespace$new(xml = NULL, prefix, uri)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{xml_node-class}}

\item{\code{prefix}}{prefix}

\item{\code{uri}}{uri}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSNamespace-decode"></a>}}
\if{latex}{\out{\hypertarget{method-GSNamespace-decode}{}}}
\subsection{Method \code{decode()}}{
Decodes from XML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSNamespace$decode(xml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{xml_node-class}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSNamespace-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GSNamespace-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSNamespace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
