% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSNamespace.R
\docType{class}
\name{GSNamespace}
\alias{GSNamespace}
\title{Geoserver REST API Namespace}
\format{\code{\link{R6Class}} object.}
\usage{
GSNamespace
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer namespace
}
\description{
Geoserver REST API Namespace
}
\section{Fields}{

\describe{
\item{\code{name}}{}

\item{\code{prefix}}{}

\item{\code{uri}}{}

\item{\code{full}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, prefix, uri)}}{
   This method is used to instantiate a GSNamespace
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a GSNamespace from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a GSNamespace to XML
 }
}
}
\examples{
GSNamespace$new(prefix = "prefix", uri = "http://prefix")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{namespace}
\keyword{rest}

