% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSRESTEntrySet.R
\docType{class}
\name{GSRESTEntrySet}
\alias{GSRESTEntrySet}
\title{Geoserver REST API XML entry set}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a entry set
}
\description{
Geoserver REST API XML entry set

Geoserver REST API XML entry set
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{entryset}
\keyword{geoserver}
\keyword{rest}
\section{Super class}{
\code{\link[geosapi:GSRESTResource]{geosapi::GSRESTResource}} -> \code{GSRESTEntrySet}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entryset}}{entryset}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GSRESTEntrySet-new}{\code{GSRESTEntrySet$new()}}
\item \href{#method-GSRESTEntrySet-decode}{\code{GSRESTEntrySet$decode()}}
\item \href{#method-GSRESTEntrySet-setEntryset}{\code{GSRESTEntrySet$setEntryset()}}
\item \href{#method-GSRESTEntrySet-addEntry}{\code{GSRESTEntrySet$addEntry()}}
\item \href{#method-GSRESTEntrySet-setEntry}{\code{GSRESTEntrySet$setEntry()}}
\item \href{#method-GSRESTEntrySet-delEntry}{\code{GSRESTEntrySet$delEntry()}}
\item \href{#method-GSRESTEntrySet-clone}{\code{GSRESTEntrySet$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSRESTResource" data-id="encode"><a href='../../geosapi/html/GSRESTResource.html#method-GSRESTResource-encode'><code>geosapi::GSRESTResource$encode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSRESTResource" data-id="getClassName"><a href='../../geosapi/html/GSRESTResource.html#method-GSRESTResource-getClassName'><code>geosapi::GSRESTResource$getClassName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="geosapi" data-topic="GSRESTResource" data-id="print"><a href='../../geosapi/html/GSRESTResource.html#method-GSRESTResource-print'><code>geosapi::GSRESTResource$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRESTEntrySet-new"></a>}}
\if{latex}{\out{\hypertarget{method-GSRESTEntrySet-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object of class \link{GSRESTEntrySet}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRESTEntrySet$new(rootName, xml = NULL, entryset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rootName}}{root name}

\item{\code{xml}}{object of class \link{xml_node-class}}

\item{\code{entryset}}{entry set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRESTEntrySet-decode"></a>}}
\if{latex}{\out{\hypertarget{method-GSRESTEntrySet-decode}{}}}
\subsection{Method \code{decode()}}{
Decodes from XML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRESTEntrySet$decode(xml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link{xml_node-class}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRESTEntrySet-setEntryset"></a>}}
\if{latex}{\out{\hypertarget{method-GSRESTEntrySet-setEntryset}{}}}
\subsection{Method \code{setEntryset()}}{
Set entry set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRESTEntrySet$setEntryset(entryset)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entryset}}{entry set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRESTEntrySet-addEntry"></a>}}
\if{latex}{\out{\hypertarget{method-GSRESTEntrySet-addEntry}{}}}
\subsection{Method \code{addEntry()}}{
Adds entry set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRESTEntrySet$addEntry(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{key}

\item{\code{value}}{value}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if added, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRESTEntrySet-setEntry"></a>}}
\if{latex}{\out{\hypertarget{method-GSRESTEntrySet-setEntry}{}}}
\subsection{Method \code{setEntry()}}{
Sets entry set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRESTEntrySet$setEntry(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{key}

\item{\code{value}}{value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRESTEntrySet-delEntry"></a>}}
\if{latex}{\out{\hypertarget{method-GSRESTEntrySet-delEntry}{}}}
\subsection{Method \code{delEntry()}}{
Deletes entry set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRESTEntrySet$delEntry(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{key}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if deleted, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRESTEntrySet-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GSRESTEntrySet-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRESTEntrySet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
