% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSRESTEntrySet.R
\docType{class}
\name{GSRESTEntrySet}
\alias{GSRESTEntrySet}
\title{Geoserver REST API XML entry set}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} for modelling a entry set
}
\description{
Geoserver REST API XML entry set
}
\section{Fields}{

\describe{
\item{\code{entryset}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml)}}{
   This method is used to instantiate a GSDataStore
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a GSRESTEntrySet from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a GSRESTEntrySet as XML
 }
 \item{\code{setEntryset(entryset)}}{
   Sets an entryset (list)
 }
 \item{\code{addEntry(key, value)}}{
   Adds an entry (key/value pair). Returns TRUE if added, FALSE otherwise
 }
 \item{\code{setEntry(key, value)}}{
   Sets an entry (key/value pair).
 }
 \item{\code{delEntry(key)}}{
   Deletes an entry by key. Returns TRUE if removed, FALSE otherwise
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{entryset}
\keyword{geoserver}
\keyword{rest}
