% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSRESTResource.R
\docType{class}
\name{GSRESTResource}
\alias{GSRESTResource}
\title{Geoserver REST API REST Resource interface}
\format{\code{\link{R6Class}} object.}
\usage{
GSRESTResource
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer REST resource interface
}
\description{
Geoserver REST API REST Resource interface
}
\section{Abstract Methods}{

\describe{
 \item{\code{new()}}{
   This method is used to instantiate a GSRESTResource
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a GSDataStore from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a GSDataStore as XML
 }
}
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{rest}

