% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSRESTResource.R
\docType{class}
\name{GSRESTResource}
\alias{GSRESTResource}
\title{Geoserver REST API REST Resource interface}
\format{
\code{\link[R6]{R6Class}} object.
}
\value{
Object of \code{\link[R6]{R6Class}} for modelling a GeoServer REST resource interface
}
\description{
Geoserver REST API REST Resource interface

Geoserver REST API REST Resource interface
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{rest}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rootName}}{root name}

\item{\code{attrs}}{attrs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GSRESTResource-new}{\code{GSRESTResource$new()}}
\item \href{#method-GSRESTResource-decode}{\code{GSRESTResource$decode()}}
\item \href{#method-GSRESTResource-encode}{\code{GSRESTResource$encode()}}
\item \href{#method-GSRESTResource-print}{\code{GSRESTResource$print()}}
\item \href{#method-GSRESTResource-getClassName}{\code{GSRESTResource$getClassName()}}
\item \href{#method-GSRESTResource-clone}{\code{GSRESTResource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRESTResource-new"></a>}}
\if{latex}{\out{\hypertarget{method-GSRESTResource-new}{}}}
\subsection{Method \code{new()}}{
Initializes an object of class \link{GSRESTResource}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRESTResource$new(xml, rootName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link[xml2]{xml_node-class}}

\item{\code{rootName}}{root name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRESTResource-decode"></a>}}
\if{latex}{\out{\hypertarget{method-GSRESTResource-decode}{}}}
\subsection{Method \code{decode()}}{
Decodes from XML. Abstract method to be implemented by sub-classes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRESTResource$decode(xml)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xml}}{object of class \link[xml2]{xml_node-class}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRESTResource-encode"></a>}}
\if{latex}{\out{\hypertarget{method-GSRESTResource-encode}{}}}
\subsection{Method \code{encode()}}{
Encodes as XML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRESTResource$encode()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \link[xml2]{xml_node-class}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRESTResource-print"></a>}}
\if{latex}{\out{\hypertarget{method-GSRESTResource-print}{}}}
\subsection{Method \code{print()}}{
Provides a custom print output (as tree) of the current class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRESTResource$print(..., depth = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{args}

\item{\code{depth}}{class nesting depth}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRESTResource-getClassName"></a>}}
\if{latex}{\out{\hypertarget{method-GSRESTResource-getClassName}{}}}
\subsection{Method \code{getClassName()}}{
Get class name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRESTResource$getClassName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an object of class \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GSRESTResource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GSRESTResource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GSRESTResource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
