% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSVirtualTableGeometry.R
\docType{class}
\name{GSVirtualTableGeometry}
\alias{GSVirtualTableGeometry}
\title{Geoserver REST API GSVirtualTableGeometry}
\format{\code{\link{R6Class}} object.}
\usage{
GSVirtualTableGeometry
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer virtual table geometry
}
\description{
Geoserver REST API GSVirtualTableGeometry
}
\section{Fields}{

\describe{
\item{\code{srid}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(xml, name, type, srid)}}{
   This method is used to instantiate a GSVirtualTableGeometry
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a GSVirtualTableGeometry from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a GSVirtualTableGeometry to XML. Inherited from the
   generic \code{GSRESTResource} encoder
 }
}
}

\examples{
GSVirtualTableGeometry$new(name = "work", type = "MultiPolygon", srid = 4326)

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{rest}
\keyword{virtualTable}
