% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSWorkspace.R
\docType{class}
\name{GSWorkspace}
\alias{GSWorkspace}
\title{Geoserver REST API Workspace}
\format{\code{\link{R6Class}} object.}
\usage{
GSWorkspace
}
\value{
Object of \code{\link{R6Class}} for modelling a GeoServer workspace
}
\description{
Geoserver REST API Workspace
}
\section{Fields}{

\describe{
\item{\code{name}}{}
}}
\section{Methods}{

\describe{
 \item{\code{new(xml, name)}}{
   This method is used to instantiate a GSWorkspace
 }
 \item{\code{decode(xml)}}{
   This method is used to decode a GSWorkspace from XML
 }
 \item{\code{encode()}}{
   This method is used to encode a GSWorkspace to XML. Inherited from the
   generic \code{GSRESTResource} encoder
 }
}
}
\examples{
GSWorkspace$new(name = "work")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{rest}
\keyword{workspace}

