% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSWorkspaceManager.R
\docType{class}
\name{GSWorkspaceManager}
\alias{GSWorkspaceManager}
\title{Geoserver REST API Manager}
\format{\code{\link{R6Class}} object.}
\usage{
GSWorkspaceManager
}
\value{
Object of \code{\link{R6Class}} with methods for managing the workspaces
 of a GeoServer instance.
}
\description{
Geoserver REST API Manager
}
\section{Methods}{

\describe{
 \item{\code{new(}}{
   This method is used to instantiate a GSWorkspaceManager
 }
 \item{\code{getWorkspaces()}}{
   Get the list of available workspace. Returns an object of class \code{xml_nodeset}
 }
 \item{\code{getWorkspaceNames()}}{
   Get the list of available workspace names. Returns an vector of class
   \code{character}
 }
 \item{\code{getWorkspace(name)}}{
   Get a \code{\link{GSWorkspace}} object given a workspace name.
 }
}
}
\examples{
GSWorkspaceManager$new("http://localhost:8080/geoserver", "admin", "geoserver")
}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{api}
\keyword{geoserver}
\keyword{rest}
\keyword{workspace}

